✨ feat: Add guide for setting up a Kubernetes cluster with RKE2 on Ubuntu 22.04
This commit adds a detailed guide for setting up a Kubernetes cluster using RKE2 on Ubuntu 22.04. It covers reasons for choosing RKE2 over RKE1, prerequisites, installation steps for the server (Control Plane) and worker nodes, as well as verification of the Kubernetes cluster post-installation. Additionally, it includes instructions for running a sample application (`nginx`) in the cluster.